+++ app/src/main/java/com/concordium/wallet/ui/account/accountdetails/TransactionMappingHelper.kt
@@ -4,7 +4,6 @@ import android.content.Context
 import com.concordium.wallet.R
 import com.concordium.wallet.data.model.RemoteTransaction
 import com.concordium.wallet.data.model.Transaction
-import com.concordium.wallet.data.model.TransactionOriginType
 import com.concordium.wallet.data.room.Account
 import com.concordium.wallet.data.room.Recipient
 import com.concordium.wallet.data.room.Transfer
@@ -38,20 +37,17 @@ class TransactionMappingHelper(
         transaction: Transaction,
         remoteTransaction: RemoteTransaction,
     ) {
-        var address: String? = null
-        val source = remoteTransaction.details.transferSource
-        val destination = remoteTransaction.details.transferDestination
-        if (source != null && destination != null) {
-            address = when (remoteTransaction.origin.type) {
-                TransactionOriginType.Self -> destination
-                TransactionOriginType.Account -> source
-                else -> null
-            }
-        }
-        if (address != null) {
-            transaction.title = getCounterpartyName(address)
-        } else {
-            transaction.title = remoteTransaction.details.description
+        transaction.title = when {
+            transaction.isOriginSelf()
+                    && transaction.toAddress != null ->
+                getCounterpartyName(transaction.toAddress)
+
+            !transaction.isOriginSelf()
+                    && transaction.fromAddress != null ->
+                getCounterpartyName(transaction.fromAddress)
+
+            else ->
+                remoteTransaction.details.description
         }
     }
 }